+++ lib/model/error.dart
